+++ src/plombery/config/__init__.py
@@ -1,0 +1,4 @@
+from plombery.config.model import Settings
+
+
+settings = Settings()