+++ src/plombery/database/models.py
@@ -2,7 +2,7 @@ from typing import List
 import datetime
 
 from fastapi.encoders import jsonable_encoder
-from pydantic import parse_obj_as
+from pydantic import TypeAdapter
 import sqlalchemy as sa
 from sqlalchemy import Column, Integer, String, DateTime
 from sqlalchemy.dialects.postgresql import JSONB
@@ -33,9 +33,11 @@ class PydanticType(sa.types.TypeDecorator):
     # Ref: https://www.postgresql.org/docs/13/datatype-json.html
     impl = sa.types.JSON
 
+    cache_ok = True
+
     def __init__(self, pydantic_type):
         super().__init__()
-        self.pydantic_type = pydantic_type
+        self.adapter = TypeAdapter(pydantic_type)
 
     def load_dialect_impl(self, dialect):
         # Use JSONB for PostgreSQL and JSON for other databases.
@@ -45,10 +47,10 @@ class PydanticType(sa.types.TypeDecorator):
             return dialect.type_descriptor(sa.JSON())
 
     def process_bind_param(self, value, dialect):
-        return jsonable_encoder(value) if value else None
+        return jsonable_encoder(value) if value is not None else None
 
     def process_result_value(self, value, dialect):
-        return parse_obj_as(self.pydantic_type, value) if value else None
+        return self.adapter.validate_python(value)
 
 
 class AwareDateTime(sa.types.TypeDecorator):
@@ -59,7 +61,7 @@ class AwareDateTime(sa.types.TypeDecorator):
 
     impl = DateTime
 
-    def process_result_value(self, value, dialect):
+    def process_result_value(self, value: datetime.datetime, dialect):
         return value.replace(tzinfo=datetime.timezone.utc)
 
 
@@ -81,7 +83,9 @@ Base.metadata.create_all(bind=engine)
 def _mark_cancelled_runs():
     db = SessionLocal()
 
-    db.query(PipelineRun).filter(PipelineRun.status == PipelineRunStatus.RUNNING.value).update(
+    db.query(PipelineRun).filter(
+        PipelineRun.status == PipelineRunStatus.RUNNING.value
+    ).update(
         dict(
             status=PipelineRunStatus.CANCELLED.value,
         )