+++ src/plombery/notifications/__init__.py
@@ -1,11 +1,14 @@
 import logging
 from typing import List
-from apprise import Apprise
+from apprise import Apprise, NotifyFormat
 import apprise
 
 from plombery.config import settings
 from plombery.database.schemas import PipelineRun
-from plombery.schemas import NotificationRule
+from plombery.notifications.helpers import get_pipeline_status_verb
+from plombery.notifications.templates import render_pipeline_run
+from plombery.pipeline.pipeline import Pipeline
+from plombery.schemas import NotificationRule, PipelineRunStatus
 
 
 logger = logging.getLogger(__name__)
@@ -18,6 +21,12 @@ PIPELINE_STATUS_TO_VERB = {
     "failed": "failed ❌",
     "cancelled": "was cancelled",
 }
+
+
+def get_message_title(
+    pipeline_name: str, pipeline_run_status: PipelineRunStatus
+) -> str:
+    return f"Pipeline {pipeline_name} {get_pipeline_status_verb(pipeline_run_status)}"
 
 
 class NotificationManager:
@@ -34,11 +43,18 @@ class NotificationManager:
             rule for rule in self.rules if pipeline_run.status in rule.pipeline_status
         ]
 
-    async def notify(self, pipeline_run: PipelineRun):
+    async def notify(self, pipeline: Pipeline, pipeline_run: PipelineRun):
         rules = self._get_applicable_rules(pipeline_run)
 
         if not rules:
             return
+
+        title = get_message_title(pipeline.name, pipeline_run.status)
+        html = render_pipeline_run(
+            pipeline.name,
+            get_pipeline_status_verb(pipeline_run.status),
+            f"{settings.frontend_url}/pipelines/{pipeline_run.pipeline_id}/triggers/{pipeline_run.trigger_id}/runs/{pipeline_run.id}",
+        )
 
         apobj = Apprise()
 
@@ -48,12 +64,9 @@ class NotificationManager:
 
         with apprise.LogCapture(level=apprise.logging.INFO) as output:
             result = await apobj.async_notify(
-                title=f"[Plombery] Your pipeline {pipeline_run.pipeline_id} {PIPELINE_STATUS_TO_VERB[pipeline_run.status]}",
-                body=f"""
-    Your pipeline {pipeline_run.pipeline_id} {PIPELINE_STATUS_TO_VERB[pipeline_run.status]}
-
-    To have more info: {settings.frontend_url}/pipelines/{pipeline_run.pipeline_id}/triggers/{pipeline_run.trigger_id}/runs/{pipeline_run.id}
-    """,
+                title=title,
+                body=html,
+                body_format=NotifyFormat.HTML,
             )
 
             if result: