+++ src/plombery/schemas.py
@@ -13,15 +13,15 @@ class PipelineRunStatus(str, Enum):
 
 
 class TaskRun(BaseModel):
-    duration: Optional[NonNegativeFloat]
+    duration: Optional[NonNegativeFloat] = 0
     """Task duration in milliseconds"""
     has_output: bool = False
     """True if the task generated an output"""
-    status: Optional[PipelineRunStatus]
+    status: Optional[PipelineRunStatus] = PipelineRunStatus.PENDING
     task_id: str
 
     class Config:
-        orm_mode = True
+        from_attributes = True
 
 
 class NotificationRule(BaseModel):